requirements.txt
python-telegram-bot[job-queue]==22.7
openai>=1.30.0
google-cloud-speech>=2.26.0
cryptography>=43.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
